refactor code, add docstring
